// src/enums/nationality.ts
export enum nationality {
  Australian = 'Australian',
  British = 'British',
  Canadian = 'Canadian',
  Chinese = 'Chinese',
  Danish = 'Danish',
  Dutch = 'Dutch',
  Finnish = 'Finnish',
  French = 'French',
  Japanese = 'Japanese',
  Monegasque = 'Monegasque',
  Mexican = 'Mexican',
  Spanish = 'Spanish',
  German = 'German',
  Thai = 'Thai',
}